Done Post add Storage and delete old screen
