+++ lib/app_layout_BM/screens/BM_opeartion_screen.dart
@@ -1,7 +1,10 @@
+import 'package:bloc_v2/add_storage/add_storage_model.dart';
+import 'package:cherry_toast/resources/arrays.dart';
 import 'package:flutter/material.dart';
 import 'package:google_fonts/google_fonts.dart';
 import 'dart:convert';
 import 'package:http/http.dart' as http;
+import 'package:cherry_toast/cherry_toast.dart';
 
 class BranchMangerOpeartion extends StatefulWidget {
   @override
@@ -40,6 +43,10 @@ class _BranchMangerOpeartionState extends State<BranchMangerOpeartion> {
     }
   }
 
+  void clearFormFields() {
+    _formKey.currentState?.reset();
+  }
+
   @override
   Widget build(BuildContext context) {
     return Scaffold(
@@ -58,9 +65,46 @@ class _BranchMangerOpeartionState extends State<BranchMangerOpeartion> {
                   functionName: 'add-storage',
                   attributeNames: const ['Storage Name', 'storageAddress', 'managerId'],
                   managerEmployees: managerEmployees,
-                  onSubmit: (values) {
-                    // TODO: Call the post function for add-storage
-                    // Example: PostFunction.addStorage(values);
+                  onSubmit: (values) async {
+                    if (_formKey.currentState!.validate()) {
+                      try {
+                        final addStorage_Model = await addStorageModel(
+                          storageName: values['Storage Name']!,
+                          storageAddress: values['storageAddress']!,
+                          managerId: values['managerId']!,
+                        );
+                        print('Adding Storage: $addStorage_Model');
+                        CherryToast.success(
+                          animationType: AnimationType.fromRight,
+                          toastPosition: Position.bottom,
+                          description: const Text(
+                            "Add Storage successfully",
+                            style: TextStyle(color: Colors.black),
+                          ),
+                        ).show(context);
+                        clearFormFields();
+                      } catch (e) {
+                        print('Error adding storage: $e');
+                        CherryToast.error(
+                          toastPosition: Position.bottom,
+                          animationType: AnimationType.fromRight,
+                          description: const Text(
+                            "Something went wrong!",
+                            style: TextStyle(color: Colors.black),
+                          ),
+                        ).show(context);
+                      }
+                    } else {
+                      print('Form is not valid');
+                      CherryToast.warning(
+                        toastPosition: Position.bottom,
+                        animationType: AnimationType.fromLeft,
+                        description: const Text(
+                          "Data is not valid or not complete",
+                          style: TextStyle(color: Colors.black),
+                        ),
+                      ).show(context);
+                    }
                   },
                 ),
                 FunctionInputTile(
@@ -164,6 +208,12 @@ class _FunctionInputTileState extends State<FunctionInputTile> {
     setState(() {
       _isExpanded = !_isExpanded;
     });
+  }
+
+  void clearForm() {
+    for (var controller in _textControllers) {
+      controller.clear();
+    }
   }
 
   @override
@@ -305,18 +355,14 @@ class _FunctionInputTileState extends State<FunctionInputTile> {
                           ),
                         ),
                       ),
-                      onPressed: () {
+                      onPressed: () async {
                         if (Form.of(context)?.validate() ?? false) {
                           final values = {
                             for (int i = 0; i < widget.attributeNames.length; i++)
                               widget.attributeNames[i]: _textControllers[i].text
                           };
-                          // Call the onSubmit function with the form values
                           widget.onSubmit(values);
-
-                          // TODO: Call the post function for each specific use case here.
-                          // For example: PostFunction.addStorage(values);
-                          // Your partner should implement the PostFunction class with appropriate methods to handle the data submission.
+                          clearForm();
                         }
                       },
                       child: Text('Submit'),
